requirements.txt
asgiref==3.8.1
Django==4.2.13
django-restframework==0.0.1
djangorestframework==3.15.2
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
